Update method of verifying requests from Slack.
The previous implementation used the verification token, which is
deprecated and may not work in the future. The new method uses a
signing secret and generates a signature for each request (to protect
against replay attacks).

Also fixes a bug where the generic "unauthenticated" page would be
rendered after the JSON returned by the bot.