Add support for Slack DM notifications for specific puzzles.
The puzzle edit page now has a checkbox that allows you to opt into
Slack notifications for puzzle updates. Currently these are triggered
whenever the status changes (e.g. stuck, priority) or the puzzle is
solved/unsolved.

We use the PuzzleMember table for this, since effectively we want to
track a many-to-many relationship between Puzzles and Members. This
table was originally used to track solvers, but that functionality has
mostly been replaced with tracking in the Google Sheet, and Propel seems
to choke a bit on having another table defined with exactly the same
schema.

See #9